requirements.txt
fastapi== 0.115.13
uvicorn==0.34.3
pandas==2.3.0
python-dotenv==1.1.0
google-generativeai==0.8.5
pydantic==2.11.7
python-multipart==0.0.20
streamlit==1.46.0
plotly==6.1.2
requests==2.32.4